backend/pyproject.toml
[tool.poetry]
name = "nine-pjt"
version = "0.0.1"
description = ""
authors = ["nine57 <gusrn015@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.3"
psycopg2-binary = "2.9.6"
djangorestframework = "3.14.0"
django-cors-headers = "4.2.0"
django-extensions = "3.2.3"
black = "23.7.0"
flake8 = "6.1.0"
requests = "2.31.0"
pillow = "10.2.0"
django-ckeditor = "6.7.0"
boto3 = "1.34.25"
django-storages = "1.14.2"
collectfast = "2.2.0"
gunicorn = "21.2.0"
pytest-django = "4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
